Commit before testing

+++ common.h
@@ -9,21 +9,20 @@
 
 using namespace std;
 
-void swap_ith_and_firstone(vector<vector<double>> &rref, vector<int> &ones, vector<int> &nonz, unsigned int &i);
+void swap_ith_and_firstone(vector<vector<double>> &rref, vector<unsigned int> &ones, vector<unsigned int> &nonz,
+                           unsigned int &i);
 
-void swap_ith_and_firstnz(vector<vector<double>> &rref, vector<int> &nonz, unsigned int &i);
+void swap_ith_and_firstnz(vector<vector<double>> &rref, vector<unsigned int> &nonz, unsigned int &i);
 
 void rowechform_piv2(vector<vector<double>> &rref, unsigned int &i, unsigned short int &n);
 
-void
-rowechform_piv(vector<vector<double>> &rref, vector<int> &nonz, unsigned int &i, unsigned short int &j, unsigned int &k,
-               unsigned short int &n);
+void rowechform_piv(vector<vector<double>> &rref, vector<unsigned int> &nonz, unsigned int &i, unsigned short int &j,
+                    unsigned int &k, unsigned short int &n);
 
-void rowechform_loop(vector<vector<double>> &rref, vector<bool> &J, unsigned int &i, unsigned short int &j,
-                     unsigned short int &rank, double &prec, unsigned short int &n);
+void rowechform_loop(vector<vector<double>> &rref, vector<bool> &J, unsigned int &i, unsigned short int &j, int &rank,
+                     double &prec, unsigned short int &n);
 
-void rowechform(vector<vector<double>> &Arref, vector<bool> &J, vector<bool> &B, unsigned short int &n,
-                unsigned short int &rank);
+void rowechform(vector<vector<double>> &Arref, vector<bool> &J, vector<bool> &B, unsigned short int &n, int &rank);
 
 bool binrank(vector<vector<double>> &Arref, vector<bool> &J, vector<bool> &b, unsigned short int &n);
